feat: saving a map to a collection supports monorepos
