+++ src/lib/appmapDir.ts
@@ -1,4 +1,4 @@
-import { readFile, writeFile } from 'fs/promises';
+import { readFile } from 'fs/promises';
 import { load } from 'js-yaml';
 import assert from 'node:assert';
 import { join } from 'path';
@@ -15,22 +15,5 @@ export async function lookupAppMapDir(folder: string): Promise<string | undefine
     } catch {
       // Unparseable AppMap config, or related error.
     }
-  }
-}
-
-export async function saveAppMapDir(folder: string, appmapDir: string): Promise<void> {
-  const appmapConfigFilePath = join(folder, 'appmap.yml');
-  if (await fileExists(appmapConfigFilePath)) {
-    let appmapConfig: unknown;
-    try {
-      appmapConfig = load(await readFile(appmapConfigFilePath, 'utf-8'));
-      assert(appmapConfig && typeof appmapConfig === 'object');
-    } catch (e) {
-      // Unparseable AppMap config, or related error.
-      console.warn(e);
-      return;
-    }
-    appmapConfig['appmap_dir'] = appmapDir;
-    await writeFile(appmapConfigFilePath, appmapDir);
   }
 }
